requirements.txt
Django>=5.0,<6.0
django-tailwind[reload]==4.0.1
Pillow>=10.0.0
qrcode>=7.4.2
python-decouple>=3.8
requests>=2.31.0
user-agents>=2.2.0
